各项删除时关联的检测

+++ tianfang/tianfang-order/src/main/java/com/tianfang/order/impl/SportMSpecValuesImpl.java
@@ -5,9 +5,13 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.tianfang.common.constants.DataStatus;
 import com.tianfang.common.util.BeanUtils;
+import com.tianfang.order.dao.SportMProductSpecValuesDao;
 import com.tianfang.order.dao.SportMSpecValuesDao;
+import com.tianfang.order.dto.SportMProductSpecValuesDto;
 import com.tianfang.order.dto.SportMSpecValuesDto;
+import com.tianfang.order.pojo.SportMProductSpecValues;
 import com.tianfang.order.pojo.SportMSpecValues;
 import com.tianfang.order.service.ISportMSpecValuesService;
 
@@ -16,6 +20,9 @@ public class SportMSpecValuesImpl implements ISportMSpecValuesService{
 
 	@Autowired
 	private SportMSpecValuesDao sportMSpecValuesDao;
+	
+	@Autowired
+	private SportMProductSpecValuesDao sportMProductSpecValuesDao;
 
 	@Override
 	public long save(SportMSpecValuesDto sportMSpecValue) {
@@ -30,13 +37,29 @@ public class SportMSpecValuesImpl implements ISportMSpecValuesService{
 	}
 
 	@Override
-	public long delete(String id) {
+	public Object delete(String id) {
 		String[] ids = id.split(",");
+		SportMSpecValuesDto sportMSpecValuesDto = new SportMSpecValuesDto();
 		long stat = 0;
 		for (String str : ids) {
+			SportMProductSpecValuesDto sportMProductSpecValues = new SportMProductSpecValuesDto();
+			sportMProductSpecValues.setSpecValuesId(str);
+			List<SportMProductSpecValues> specValues = sportMProductSpecValuesDao.findProductSpecValues(sportMProductSpecValues);
+			SportMSpecValues sportMSpecValues = sportMSpecValuesDao.selectByPrimaryKey(str);
+			if (specValues.size()>0) {
+				sportMSpecValuesDto = BeanUtils.createBeanByTarget(sportMSpecValues, SportMSpecValuesDto.class);
+				sportMSpecValuesDto.setDeleteStat(DataStatus.DISABLED);
+				return sportMSpecValuesDto;
+			}
 			stat = sportMSpecValuesDao.delete(str);
 		}
-		return stat;
+		for (String str : ids) {
+			SportMSpecValues sportMSpecValues = sportMSpecValuesDao.selectByPrimaryKey(str);
+			sportMSpecValues.setStat(DataStatus.ENABLED);
+			sportMSpecValuesDao.updateByPrimaryKeySelective(sportMSpecValues);			
+		}
+		sportMSpecValuesDto.setDeleteStat(DataStatus.ENABLED);
+		return sportMSpecValuesDto;
 	}
 
 	@Override
